backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "uxpulse-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "sqlalchemy>=2.0",
  "psycopg[binary]>=3.1",
  "pydantic>=2.6",
  "python-dotenv>=1.0",
  "openai>=1.40.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
